move diff_hashmap to diff mod
