+++ src/diff/mod.rs
@@ -1,3 +1,4 @@
+use std::collections::{HashMap, HashSet};
 use termcolor::Color;
 
 pub mod entry;
@@ -100,4 +101,61 @@ where
 
         Ok(())
     }
+}
+
+/// Compare to HashMaps of the same value type with each other, returning a bool indicating whether
+/// there are any differences and a Vec<DiffResult<A>> listing all differences
+pub fn diff_hashmap<'a, A>(
+    a: &'a HashMap<String, A>,
+    b: &'a HashMap<String, A>,
+) -> (bool, Vec<DiffResult<'a, A>>)
+where
+    A: Diff,
+{
+    let mut keys = HashSet::new();
+    keys.extend(a.keys());
+    keys.extend(b.keys());
+
+    let mut keys: Vec<_> = keys.iter().collect();
+    keys.sort();
+
+    let mut acc: Vec<DiffResult<A>> = Vec::new();
+
+    let mut has_differences = false;
+
+    for key in keys {
+        let el_a: Option<&A> = a.get(*key);
+        let el_b: Option<&A> = b.get(*key);
+
+        match (el_a, el_b) {
+            // both a and b have the key
+            (Some(v_a), Some(v_b)) => {
+                let dr: DiffResult<A> = v_a.diff(v_b);
+
+                if let DiffResult::Identical { .. } = dr {
+                } else {
+                    has_differences = true;
+                }
+
+                acc.push(dr);
+            }
+
+            // only a has the key
+            (Some(v_a), None) => {
+                has_differences = true;
+                acc.push(DiffResult::OnlyLeft { left: v_a })
+            }
+
+            // only b has the key
+            (None, Some(v_b)) => {
+                has_differences = true;
+                acc.push(DiffResult::OnlyRight { right: v_b })
+            }
+
+            // none have the key (this shouldn't happen)
+            (None, None) => {}
+        }
+    }
+
+    (has_differences, acc)
 }
